Remove github.event from input parameters
This was copied in from an example using manual dispatch only. In that
case, the only difference is appearently that it converts bools to strings.
But for the workflow to also function when called with parameters,
the github.event must not be there. Which then also means bools are bools.