Allow comparison of SpectralType directly to string
If a string represents a valid constructor, it can be compare directly to
an instance of SpectralType, which can in some cases be a lot more
convenient than having to manually convert the string to a SpectralType.

Implementing the gt and ge methods was necessary to also allow reverse
comparisons, i.e. `"B0V" < SpectralType("A0V")`.